requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1